V-2.40.04 NC-processor PRC_cut1 V-1.1: new function TCPOS (toolChangePosition) MS-Win display-html-bug fixed

+++ src/ut/ut_mem.h
@@ -93,7 +93,7 @@ void* MEM_alloc_tmp (int);
 #else
 #define MEM_alloc_tmp(siz)\
  alloca(siz);\
- if(siz >= SPC_MAX_STK) printf("*** ERR MEM_alloc_tmp |%s| %d\n",__func__,siz);\
+ if(siz >= SPC_MAX_STK) printf("*** ERR MEM_alloc_tmp |%s| %d\n",__func__,(int)siz);\
  fflush(stdout)
 #endif
